IS-1682: Move create-method for ReceivedDialogmelding
