+++ src/main/kotlin/no/nav/syfo/application/DialogmeldingProcessor.kt
@@ -3,7 +3,6 @@ package no.nav.syfo.application
 import io.ktor.client.*
 import net.logstash.logback.argument.StructuredArguments
 import no.nav.helse.dialogmelding.XMLDialogmelding
-import no.nav.helse.eiFellesformat2.XMLEIFellesformat
 import no.nav.helse.eiFellesformat2.XMLMottakenhetBlokk
 import no.nav.helse.msgHead.XMLMsgHead
 import no.nav.syfo.Environment
@@ -25,7 +24,6 @@ import no.nav.syfo.services.*
 import no.nav.syfo.util.*
 import no.nav.syfo.validation.isKodeverkValid
 import java.time.Duration
-import java.time.ZoneId
 
 class DialogmeldingProcessor(
     val database: DatabaseInterface,
@@ -107,7 +105,7 @@ class DialogmeldingProcessor(
         )
         val starttime = System.currentTimeMillis()
 
-        val receivedDialogmelding = createReceivedDialogmelding(
+        val receivedDialogmelding = ReceivedDialogmelding.create(
             dialogmeldingId = dialogmeldingId,
             fellesformat = fellesformat,
             inputMessageText = inputMessageText,
@@ -202,53 +200,6 @@ class DialogmeldingProcessor(
             ),
             StructuredArguments.keyValue("latency", duration.toMillis()),
             StructuredArguments.fields(loggingMeta)
-        )
-    }
-
-    fun createReceivedDialogmelding(
-        dialogmeldingId: String,
-        fellesformat: XMLEIFellesformat,
-        inputMessageText: String,
-    ): ReceivedDialogmelding {
-        val emottakblokk = fellesformat.get<XMLMottakenhetBlokk>()
-        val msgHead: XMLMsgHead = fellesformat.get()
-        val legeIdent = emottakblokk.avsenderFnrFraDigSignatur
-        val legekontorOrgName = extractSenderOrganisationName(fellesformat)
-        val legekontorHerId = extractOrganisationHerNumberFromSender(fellesformat)?.id
-        val dialogmeldingXml = extractDialogmelding(fellesformat)
-        val dialogmeldingType = findDialogmeldingType(emottakblokk.ebService, emottakblokk.ebAction)
-        val legeHpr = extractLegeHpr(dialogmeldingId, fellesformat)
-        val behandlerNavn = extractBehandlerNavn(fellesformat)
-        val behandlerIdent = extractIdentFromBehandler(fellesformat)
-        val innbyggerIdent = extractInnbyggerident(fellesformat)
-        val legekontorOrgNr = extractOrganisationNumberFromSender(fellesformat)?.id
-
-        if (behandlerIdent != legeIdent) {
-            logger.info("Behandler and avsender are different in dialogmelding: $dialogmeldingId")
-        }
-
-        val dialogmelding = dialogmeldingXml.toDialogmelding(
-            dialogmeldingId = dialogmeldingId,
-            dialogmeldingType = dialogmeldingType,
-            signaturDato = msgHead.msgInfo.genDate,
-            navnHelsePersonellNavn = behandlerNavn
-        )
-
-        return ReceivedDialogmelding(
-            dialogmelding = dialogmelding,
-            personNrPasient = innbyggerIdent!!,
-            personNrLege = legeIdent,
-            navLogId = emottakblokk.ediLoggId,
-            msgId = msgHead.msgInfo.msgId,
-            legekontorOrgNr = legekontorOrgNr,
-            legekontorOrgName = legekontorOrgName,
-            legekontorHerId = legekontorHerId,
-            mottattDato = emottakblokk.mottattDatotid.toGregorianCalendar().toZonedDateTime()
-                .withZoneSameInstant(
-                    ZoneId.of("Europe/Oslo")
-                ).toLocalDateTime(),
-            legehpr = legeHpr,
-            fellesformat = inputMessageText,
         )
     }
